Chaque composant ui dans son fichier et dans le répertoire ui
