+++ src/modules/chaleur-renouvelable/client/results/ui/NoResultSection.tsx
