+++ src/modules/chaleur-renouvelable/client/results/ui/Stars.tsx
@@ -1,0 +1,18 @@
+import Image from '@/components/ui/Image';
+import Tooltip from '@/components/ui/Tooltip';
+
+export function Stars({ value }: { value: number }) {
+  return (
+    <div className="flex items-center gap-1" aria-label={`${value} étoiles`}>
+      {Array.from({ length: value }).map((_, index) => (
+        <Image key={index} src="/icons/icon-star.png" alt="" aria-hidden="true" width="20" height="20" />
+      ))}
+      <Tooltip
+        iconProps={{
+          className: 'text-blue',
+        }}
+        title="Classement Ademe ENR Choix"
+      />
+    </div>
+  );
+}
